+++ contracts/locked-astroport-vault/src/contract.rs
@@ -1,9 +1,11 @@
+use apollo_cw_asset::{Asset, AssetInfo};
+use apollo_utils::assets::receive_asset;
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 
 use cosmwasm_std::{
     to_json_binary, Binary, CosmosMsg, Deps, DepsMut, Env, Event, MessageInfo, QueryRequest, Reply,
-    Response, StdError, StdResult, SubMsg, Uint128, WasmQuery,
+    Response, StdResult, SubMsg, Uint128, WasmQuery,
 };
 use cw2::ensure_from_older_version;
 use cw_dex_astroport::{astroport, AstroportPool, AstroportStaking};
@@ -13,7 +15,6 @@ use cw_vault_standard::extensions::lockup::LockupExecuteMsg;
 use osmosis_std::types::osmosis::tokenfactory::v1beta1::MsgCreateDenom;
 
 use crate::error::{ContractError, ContractResponse};
-use crate::execute;
 use crate::helpers::{self, IntoInternalCall};
 use crate::msg::{
     ApolloExtensionExecuteMsg, ApolloExtensionQueryMsg, ExecuteMsg, ExtensionExecuteMsg,
@@ -27,6 +28,7 @@ use crate::state::{
     ConfigUnchecked, VaultState, BASE_TOKEN, CONFIG, FORCE_WITHDRAW_WHITELIST, POOL, STAKING,
     STATE, VAULT_TOKEN_DENOM,
 };
+use crate::{execute, query};
 
 pub const CONTRACT_NAME: &str = env!("CARGO_PKG_NAME");
 pub const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -47,13 +49,16 @@ pub fn instantiate(
     // Query pair info from astroport pair
     let pair_info = deps
         .querier
-        .query::<astroport::asset::PairInfo>(&QueryRequest::Wasm(WasmQuery::Smart {
+        .query::<astroport_v5::asset::PairInfo>(&QueryRequest::Wasm(WasmQuery::Smart {
             contract_addr: msg.pool_addr.clone(),
             msg: to_json_binary(&astroport::pair::QueryMsg::Pair {})?,
         }))?;
 
     // Store pool info
-    let liquidity_manager = deps.api.addr_validate(&msg.astroport_liquidity_manager)?;
+    let liquidity_manager = msg
+        .astroport_liquidity_manager
+        .map(|x| deps.api.addr_validate(&x))
+        .transpose()?;
     let pool = AstroportPool::new(
         deps.as_ref(),
         deps.api.addr_validate(&msg.pool_addr)?,
@@ -81,7 +86,9 @@ pub fn instantiate(
         "factory/{}/{}",
         env.contract.address, msg.vault_token_subdenom
     );
-    BASE_TOKEN.save(deps.storage, &pair_info.liquidity_token)?;
+    let base_token = AssetInfo::from_str(deps.api, &pair_info.liquidity_token);
+
+    BASE_TOKEN.save(deps.storage, &base_token)?;
     VAULT_TOKEN_DENOM.save(deps.storage, &vault_token_denom)?;
     STATE.save(
         deps.storage,
@@ -93,7 +100,7 @@ pub fn instantiate(
 
     // Store staking info
     let staking = AstroportStaking {
-        lp_token_addr: pair_info.liquidity_token,
+        lp_token: base_token.clone(),
         incentives: deps.api.addr_validate(&msg.astroport_incentives_addr)?,
     };
     STAKING.save(deps.storage, &staking)?;
@@ -108,6 +115,7 @@ pub fn instantiate(
     Ok(Response::new().add_message(create_denom_msg))
 }
 
+#[allow(deprecated)]
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn execute(
     deps: DepsMut,
@@ -117,22 +125,25 @@ pub fn execute(
 ) -> Result<Response, ContractError> {
     match msg {
         ExecuteMsg::Deposit { amount, recipient } => {
+            let base_token = BASE_TOKEN.load(deps.storage)?;
+            let deposited = Asset::new(base_token.clone(), amount);
+            let transfer_res = receive_asset(&info, &env, &deposited)?;
+
             // Call contract itself first to compound, but as a SubMsg so that we can still
             // deposit if the compound fails
             let compound_msg = SubMsg::reply_on_error(
-                ApolloExtensionExecuteMsg::Compound {}.into_internal_call(&env, vec![])?,
+                InternalMsg::Compound {
+                    discount_deposit: amount,
+                }
+                .into_internal_call(&env, vec![])?,
                 COMPOUND_REPLY_ID,
             );
 
             let recipient = helpers::unwrap_recipient(recipient, &info, deps.api)?;
-            let deposit_msg = InternalMsg::Deposit {
-                amount,
-                depositor: info.sender,
-                recipient,
-            }
-            .into_internal_call(&env, vec![])?;
+            let deposit_msg =
+                InternalMsg::Deposit { amount, recipient }.into_internal_call(&env, vec![])?;
 
-            Ok(Response::new()
+            Ok(transfer_res
                 .add_submessage(compound_msg)
                 .add_message(deposit_msg))
         }
@@ -205,18 +216,19 @@ pub fn execute(
                 }
 
                 match msg {
+                    InternalMsg::Compound { discount_deposit } => {
+                        execute::compound::execute_compound(deps, env, discount_deposit)
+                    }
                     InternalMsg::SellTokens {} => execute::compound::execute_sell_tokens(deps, env),
                     InternalMsg::ProvideLiquidity {} => {
                         execute::compound::execute_provide_liquidity(deps, env)
                     }
-                    InternalMsg::StakeLps {} => execute::compound::execute_stake_lps(deps, env),
-                    InternalMsg::Deposit {
-                        amount,
-                        depositor,
-                        recipient,
-                    } => execute::basic_vault::execute_deposit(
-                        deps, env, amount, depositor, recipient,
-                    ),
+                    InternalMsg::StakeLps { discount_tokens } => {
+                        execute::compound::execute_stake_lps(deps, env, discount_tokens)
+                    }
+                    InternalMsg::Deposit { amount, recipient } => {
+                        execute::basic_vault::execute_deposit(deps, env, amount, recipient)
+                    }
                     InternalMsg::Redeem { recipient, amount } => {
                         execute::basic_vault::execute_redeem(
                             deps, env, info, amount, recipient, false,
@@ -234,7 +246,7 @@ pub fn execute(
                     execute::basic_vault::execute_update_config(deps, info, updates)
                 }
                 ApolloExtensionExecuteMsg::Compound {} => {
-                    execute::compound::execute_compound(deps, env)
+                    execute::compound::execute_compound(deps, env, Uint128::zero())
                 }
             },
         },
@@ -246,7 +258,9 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
     match msg {
         QueryMsg::VaultStandardInfo {} => to_json_binary(&query_vault_standard_info(deps)?),
         QueryMsg::Info {} => to_json_binary(&query_vault_info(deps)?),
+        #[allow(deprecated)]
         QueryMsg::PreviewDeposit { .. } => unimplemented!("Cannot reliably preview deposit"),
+        #[allow(deprecated)]
         QueryMsg::PreviewRedeem { .. } => unimplemented!("Cannot reliably preview redeem"),
         QueryMsg::TotalAssets {} => {
             let state = STATE.load(deps.storage)?;
@@ -261,6 +275,9 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
         }
         QueryMsg::ConvertToAssets { amount } => {
             to_json_binary(&helpers::convert_to_assets(deps, amount))
+        }
+        QueryMsg::VaultTokenExchangeRate { quote_denom } => {
+            to_json_binary(&query::vault_token_exchange_rate(deps, quote_denom)?)
         }
         QueryMsg::VaultExtension(ext_msg) => match ext_msg {
             ExtensionQueryMsg::Lockup(lockup_msg) => match lockup_msg {
@@ -318,34 +335,10 @@ pub fn reply(_deps: DepsMut, _env: Env, msg: Reply) -> Result<Response, Contract
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
-pub fn migrate(mut deps: DepsMut, env: Env, msg: MigrateMsg) -> Result<Response, ContractError> {
-    let incentives_contract = deps.api.addr_validate(&msg.incentives_contract)?;
-
-    let old_version = ensure_from_older_version(deps.storage, CONTRACT_NAME, CONTRACT_VERSION)?;
-
-    let res = match old_version.to_string().as_str() {
-        "0.2.0" => {
-            crate::migrations::migrate_from_0_2_0_to_0_3_0(deps.branch())?;
-            crate::migrations::migrate_from_0_3_0_to_current(
-                deps.branch(),
-                env,
-                incentives_contract,
-            )?
-        }
-        "0.3.0" => crate::migrations::migrate_from_0_3_0_to_current(
-            deps.branch(),
-            env,
-            incentives_contract,
-        )?,
-        "0.4.0" | "0.4.1" | "0.4.2" => Response::default(),
-        _ => {
-            return Err(StdError::generic_err(
-                "Cannot migrate from a version of the contract other than v0.2.0, v0.3.0, v0.4.0, or v0.4.1",
-            ).into())
-        }
-    };
+pub fn migrate(deps: DepsMut, _env: Env, _msg: MigrateMsg) -> Result<Response, ContractError> {
+    ensure_from_older_version(deps.storage, CONTRACT_NAME, CONTRACT_VERSION)?;
 
     cw2::set_contract_version(deps.storage, CONTRACT_NAME, CONTRACT_VERSION)?;
 
-    Ok(res)
+    Ok(Response::new())
 }