+++ contracts/locked-astroport-vault/src/execute/lockup.rs
@@ -5,7 +5,7 @@ use crate::error::{ContractError, ContractResponse};
 use crate::helpers::unwrap_recipient;
 use crate::state::{self, BASE_TOKEN, FORCE_WITHDRAW_WHITELIST, STAKING};
 
-use cw_dex::traits::Unstake;
+use cw_dex_astroport::cw_dex::traits::Unstake;
 
 pub fn execute_withdraw_unlocked(
     deps: DepsMut,
@@ -25,7 +25,7 @@ pub fn execute_withdraw_unlocked(
     let res = staking.unstake(deps.as_ref(), &env, claim_amount)?;
 
     // Send LP tokens to recipient
-    let send_msg = Asset::cw20(base_token, claim_amount).transfer_msg(recipient)?;
+    let send_msg = Asset::new(base_token, claim_amount).transfer_msg(recipient)?;
 
     let event = Event::new("apollo/vaults/execute_withdraw_unlocked")
         .add_attribute("lockup_id", format!("{}", lockup_id))
@@ -95,7 +95,7 @@ pub fn execute_force_withdraw_unlocking(
     let res = staking.unstake(deps.as_ref(), &env, claim_amount)?;
 
     // Send LP tokens to recipient
-    let send_msg = Asset::cw20(base_token, claim_amount).transfer_msg(recipient)?;
+    let send_msg = Asset::new(base_token, claim_amount).transfer_msg(recipient)?;
 
     let event = Event::new("apollo/vaults/execute_force_withdraw_unlocking")
         .add_attribute("lockup_id", format!("{}", lockup_id))